Add too complicated "series" handling to book pages
Sometimes we want to call something "The FOO series" and sometimes "The"
is already part of the name. So handle that.

Also sometimes a series name already has a "series-like" word in it.
saga, trilogy, etc. We don't want to call something "The FOO SAGA
series". So also handle that.